requirements.txt
# Virtual Env
virtualenv

# Core
django
djangorestframework
Pillow
requests

# Auth Token
djangorestframework-jwt
